+++ src/Sulu/Bundle/HttpCacheBundle/DependencyInjection/Configuration.php
@@ -64,7 +64,9 @@ class Configuration implements ConfigurationInterface
                             ->fixXmlConfig('server')
                             ->children()
                                 ->arrayNode('servers')
-                                    ->beforeNormalization()->ifString()->then(function ($v) { return preg_split('/\s*,\s*/', $v); })->end()
+                                    ->beforeNormalization()->ifString()->then(function ($v) {
+                                        return preg_split('/\s*,\s*/', $v);
+                                    })->end()
                                     ->useAttributeAsKey('name')
                                     ->isRequired()
                                     ->requiresAtLeastOneElement()