+++ src/Sulu/Component/Rest/Tests/Unit/DoctrineRestHelperTest.php
@@ -45,6 +45,7 @@ class DoctrineRestHelperTest extends \PHPUnit_Framework_TestCase
         $entities->expects($this->once())->method('getIterator')->willReturn(new \ArrayIterator([null, null]));
         $entities->expects($this->exactly(2))->method('add');
 
-        $this->restHelper->processSubEntities($entities, [], function () {});
+        $this->restHelper->processSubEntities($entities, [], function () {
+        });
     }
 }